barakah-enterprise/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
reportlab==4.0.7
openai==1.3.7
stripe==7.8.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8
httpx==0.25.2
arabic-reshaper==3.0.0
python-bidi==0.4.2
pillow==10.1.0
qrcode==7.4.2